+++ lib/generators/shopify_app/home_controller/templates/home_controller.rb
@@ -1,4 +1,6 @@
-class HomeController < ShopifyApp::AuthenticatedController
+class HomeController < ApplicationController
+  include ShopifyApp::AuthenticatedByShopify
+
   def index
     @products = ShopifyAPI::Product.find(:all, params: { limit: 10 })
     @webhooks = ShopifyAPI::Webhook.find(:all)